Refactored packages.py into a PackageManager class. Not expecting any change in functionality.
